% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get.env.cdiscstandards.R
\name{Get.env.cdiscstandards}
\alias{Get.env.cdiscstandards}
\title{Get rdf.cdisc.org rrdf model from local file and store in environment}
\usage{
Get.env.cdiscstandards()
}
\value{
rdf.cdisc.org rrdf model
}
\description{
Get rdf.cdisc.org rrdf model from local file and store in environment
}

