% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get.filenames.for.cdisc.standards.R
\name{Get.filenames.for.cdisc.standards}
\alias{Get.filenames.for.cdisc.standards}
\title{List of files consisting of the rdf.cdisc.org standards}
\usage{
Get.filenames.for.cdisc.standards()
}
\value{
Character vector with filenames for cdisc.rdf.org standards
}
\description{
List of files consisting of the rdf.cdisc.org standards
}

