% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Load.cdisc.standards.R
\name{Load.cdisc.standards}
\alias{Load.cdisc.standards}
\title{Load CDISC standards}
\usage{

  Load.cdisc.standards(cdisc.load.zip = file.path(system.file("extdata/CDISC-standards-rdf",
  package = "rrdfcdisc"), "cdisc-rdf.zip"),
  cdisc.load.ttl = file.path(tempdir(), "cdisc-rdf.ttl"),
  remove.cdisc.load.ttl = TRUE)
}
\arguments{
\item{cdisc.load.zip}{Path and filename to zip file with turtle
version of rdf.cdisc.org. Default location is in package
extdata/CDISC-standards-rdf directory files cdisc-rdf.zip}

\item{cdisc.load.ttl}{Path and filename for the unzipped turtle
version of rdf.cdisc.org. Default location is file cdisc-rdf.ttl
in R session temporary directory}

\item{remove.cdisc.load.ttl}{If TRUE then the remove the unzipped
turtle version of rdf.cdisc.org}
}
\value{
RRDF model containing CDISC standards stored locally in
zip file
}
\description{
Load CDISC standards
}
\examples{
cubestore<- Load.cdisc.standards()
 
}

