% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrdfcdisc.R
\docType{data}
\name{qbCDISCprefixes}
\alias{qbCDISCprefixes}
\title{Default PREFIXes for SPARQL queries}
\format{A list of PREFIXes for SPARQL queries}
\usage{
data(qbCDISCprefixes)
}
\description{
Default PREFIXes for SPARQL queries
}
\keyword{datasets}

