% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.rdf.noprint.R
\name{summarize.rdf.noprint}
\alias{summarize.rdf.noprint}
\title{Return number of triples in model}
\usage{
summarize.rdf.noprint(model)
}
\arguments{
\item{model}{RRDF model}
}
\value{
Number of triples in model
}
\description{
Return number of triples in model
}
\details{
This the rrdf summarize.rdf without a print statement
}

