% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GetCDISCCodeListSparqlQuery.R
\name{GetCDISCCodeListSparqlQuery}
\alias{GetCDISCCodeListSparqlQuery}
\title{SPARQL query for CDISC codelists}
\usage{
GetCDISCCodeListSparqlQuery(forsparqlprefix, nciDomainValue)
}
\arguments{
\item{forsparqlprefix}{SPARQL prefix}

\item{nciDomainValue}{NCI domain value}
}
\value{
SPARQL query
}
\description{
SPARQL query for CDISC codelists
}

